Code refactor and documentation updated
* Merged VoucherChecker and VoucherConfirmer in a single VoucherService
* Update README
