+++ README.md
@@ -13,14 +13,21 @@ This sdk can be used by merchants to:
 The sdk has no external dependencies.
 
 ## Usage
+These are the operations that can be done usign the sdk:
 
-To check the validity and the residual amount of a given voucher create a `VoucherChecker`, invoke the `check` method passing the voucher code and finally get the amount with `getImporto()`
-    new VoucherChecker().check("XXX").getImporto();
+* Voucher check
+To check the validity and the residual amount of a given voucher create a `VoucherService`, invoke the `check` method passing the voucher code and finally get the amount with `getImporto()`
+    new VoucherService().check("XXXXXXXX").getImporto();
 
-To consume a specific amount from a given voucher create a `VoucherConfirmer`, invoke the `confirm` method passing the voucher code and the amount and finally get the result of the operation with `getEsito()`
-    new VoucherConfirmer().consume("y75Wx7xj", 0.5).getEsito()
+* Voucher consume
+To consume a specific amount from a given voucher create a `VoucherService`, invoke the `confirm` method passing the voucher code and the amount and finally get the result of the operation with `getEsito()`
+    new VoucherService().consume("XXXXXXXX", 0.5).getEsito()
 
-### Debug
+* Voucher invalidate
+To consume a specific amount from a given voucher create a `VoucherService`, invoke the `confirm` method passing the voucher code and the amount and finally get the result of the operation with `getEsito()`
+    new VoucherService().invalidate("XXXXXXXX").getImporto()
+
+## Debug
 To use the sdk with the staging environment, you need to properly set the server certificate.
 To do this you should:
 * Get the server certificate from
@@ -36,9 +43,20 @@ To do this you should:
     -Djavax.net.ssl.trustStore=wstest18appitaliait.jks
     -Djavax.net.ssl.trustStorePassword=123456
 
+## Error codes
+Those are the error codes returned by the server:
+Code | Description
+-----|------------
+`01` | Error in the input parameters, check and try again.
+`02` | The requested voucher is not available on the system. It could be alreadycollected or canceled.
+`03` | Impossible to activate the user. Please verify input parameters and that the userhas not been already activated.
+`04` | The amount claimed is greater than the amount of the selected voucher.
+`05` | User inactive, voucher impossible to verify.
+`06` | Category and type of this voucher are not aligned with category and typemanaged by the user.
+
 
 ## Test
 
-_¯\\_(ツ)_/¯_ for now
+TODO _¯\\_(ツ)_/¯_
