+++ src/it/italia/developers/Main.java
@@ -5,7 +5,7 @@ public class Main {
 	public static void main(String[] args) {
 		
 		// check
-		System.out.println(new VoucherService().check("cZd7OxFU").getImporto());
+		System.out.println(new VoucherService().check("XXXXXXXX").getImporto());
 		
 		// consume
 		System.out.println(new VoucherService().consume("XXXXXXXX", 13.37).getEsito());
